Reorganized course_structures app
Celery requires all tasks to be in a tasks module.
